+++ src/main/java/com/cmacgm/gbs/rst/api/tms/api/dto/StartTmsSessionRequest.java
@@ -1,5 +1,6 @@
 package com.cmacgm.gbs.rst.api.tms.api.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import jakarta.validation.constraints.NotNull;
@@ -9,7 +10,7 @@ import jakarta.validation.constraints.Size;
 public record StartTmsSessionRequest(
         @NotNull UUID toolkitId,
         UUID subtaskId,
-        @Positive Integer processedVolume,
+        @Positive BigDecimal processedVolume,
         @Size(max = 100) String reference,
         @Size(max = 500) String remarks) {
 }
