+++ src/main/java/com/cmacgm/gbs/rst/api/tms/api/dto/TmsSummaryResponse.java
@@ -1,7 +1,9 @@
 package com.cmacgm.gbs.rst.api.tms.api.dto;
 
+import java.math.BigDecimal;
+
 public record TmsSummaryResponse(
         long sessionsToday,
-        long totalVolume,
+        BigDecimal totalVolume,
         long pausedSessions) {
 }
